Terraform rework: modules are now run independently, each with its state.
Being able to destroy or reapply independent parts of the infra makes sense (i.e. only the downloaded images, or the networking, or some VMs).
Also, the separate states seem more secure.
The logic is, for each env, we get one separate instantiation of each module.
Tricked Terraform with symbolic links to avoid duplicating the provider, and be a little more DRY.

Still a couple of challenges to overcome:
- Can't use outputs to communicate between modules anymore since the
  state is not shared (look into local data source?)
- Can't apply all with just one command anymore. Maybe it's not such a
  big problem, because by definition the downloads and networking setup
  won't need to be applied with the same lifecycle as the rest. Worst
  case scenario, a simple script can be made to chain the apply
  commands.

The alternative in a simple, and maybe more "traditional" way would be
to have one tfvars file for each env inside a big parent module.
Simpler, but less SRP, and you always have to destroy/apply everything
together (unless you use --target).